Turn on and resolve warnings about trivial numeric casts

+++ src/metrics/counter.rs
@@ -27,7 +27,7 @@ impl Counter for StdCounter {
     // inc(double v): Increment the counter by the given amount. MUST check that v >= 0.
     // We crash with interger overflow
     fn add(&mut self, value: f64) {
-        self.value = self.value + value as f64;
+        self.value = self.value + value;
     }
 
     fn snapshot(self) -> StdCounter {
